# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swat Flood Early-Warning System API",
    "description": "MVP demo backend - temporary implementation for university evaluation",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/zones/nearest": {
      "get": {
        "tags": [
          "zones"
        ],
        "summary": "Get Nearest Zone",
        "description": "Find the nearest zone to given coordinates using PostGIS.\n\nUsed for mobile geolocation.\n\nTODO: Move to use-cases layer, add proper error handling.",
        "operationId": "get_nearest_zone_api_zones_nearest_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Nearest Zone Api Zones Nearest Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/zones/{zone_id}": {
      "get": {
        "tags": [
          "zones"
        ],
        "summary": "Get Zone",
        "description": "Get detailed information for a specific zone.",
        "operationId": "get_zone_api_zones__zone_id__get",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Zone Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZoneDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/zones/{zone_id}/risk-trend": {
      "get": {
        "tags": [
          "zones"
        ],
        "summary": "Get Zone Risk Trend",
        "description": "Get time-ordered list of risk assessments within the requested window.\n\nUsed for dashboard's risk-trend chart.\n\nTODO: Move to use-cases layer, add proper error handling.",
        "operationId": "get_zone_risk_trend_api_zones__zone_id__risk_trend_get",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Zone Id"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "24h",
              "title": "Window"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Zone Risk Trend Api Zones  Zone Id  Risk Trend Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/zones/{zone_id}/rainfall-vs-risk": {
      "get": {
        "tags": [
          "zones"
        ],
        "summary": "Get Zone Rainfall Vs Risk",
        "description": "Get paired time-ordered list of rainfall and risk data within the requested window.\n\nUsed for dashboard's rainfall-vs-risk chart.\n\nTODO: Move to use-cases layer, add proper error handling.",
        "operationId": "get_zone_rainfall_vs_risk_api_zones__zone_id__rainfall_vs_risk_get",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Zone Id"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "24h",
              "title": "Window"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Zone Rainfall Vs Risk Api Zones  Zone Id  Rainfall Vs Risk Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoordinatesSchema": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "CoordinatesSchema",
        "description": "Coordinates schema for HTTP response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RainfallWindowSchema": {
        "properties": {
          "millimetres": {
            "type": "number",
            "title": "Millimetres"
          },
          "duration_hours": {
            "type": "number",
            "title": "Duration Hours"
          }
        },
        "type": "object",
        "required": [
          "millimetres",
          "duration_hours"
        ],
        "title": "RainfallWindowSchema",
        "description": "Rainfall window schema for HTTP response."
      },
      "RiskAssessmentSchema": {
        "properties": {
          "tier": {
            "type": "string",
            "title": "Tier",
            "description": "The risk tier (SAFE, LOW, MEDIUM, HIGH, DANGER)."
          },
          "probability": {
            "type": "number",
            "title": "Probability",
            "description": "Confidence score [0, 1] of the assessment."
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "description": "Human-readable explanation of the risk."
          },
          "computed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Computed At",
            "description": "When this assessment was generated."
          },
          "combined_rain_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Combined Rain Mm",
            "description": "Combined 10-day rainfall (past 7d + forecast 72h)."
          },
          "expected_rain_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Rain Mm",
            "description": "Expected 10-day baseline rainfall."
          },
          "rainfall_anomaly_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rainfall Anomaly Ratio",
            "description": "Ratio of actual vs expected rainfall."
          },
          "top_contributing_features": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Contributing Features",
            "description": "Top SHAP contributing features from the ML model."
          }
        },
        "type": "object",
        "required": [
          "tier",
          "probability",
          "explanation",
          "computed_at"
        ],
        "title": "RiskAssessmentSchema",
        "description": "API schema for a zone risk assessment."
      },
...
      "SensorReadingSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "zone_id": {
            "type": "string",
            "title": "Zone Id"
          },
          "water_level": {
            "$ref": "#/components/schemas/WaterLevelSchema"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "zone_id",
          "water_level",
          "source",
          "timestamp"
        ],
        "title": "SensorReadingSchema",
        "description": "Sensor reading schema for HTTP response."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WaterLevelSchema": {
        "properties": {
          "metres": {
            "type": "number",
            "title": "Metres"
          }
        },
        "type": "object",
        "required": [
          "metres"
        ],
        "title": "WaterLevelSchema",
        "description": "Water level schema for HTTP response."
      },
      "WeatherSnapshotSchema": {
        "properties": {
          "zone_id": {
            "type": "string",
            "title": "Zone Id"
          },
          "rainfall": {
            "$ref": "#/components/schemas/RainfallWindowSchema"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "ttl_seconds": {
            "type": "integer",
            "title": "Ttl Seconds"
          }
        },
        "type": "object",
        "required": [
          "zone_id",
          "rainfall",
          "fetched_at",
          "ttl_seconds"
        ],
        "title": "WeatherSnapshotSchema",
        "description": "Weather snapshot schema for HTTP response."
      },
      "ZoneDetailResponse": {
        "properties": {
          "zone": {
            "$ref": "#/components/schemas/ZoneListResponse"
          },
          "recent_readings": {
            "items": {
              "$ref": "#/components/schemas/SensorReadingSchema"
            },
            "type": "array",
            "title": "Recent Readings"
          },
          "risk_history": {
            "items": {
              "$ref": "#/components/schemas/RiskAssessmentSchema"
            },
            "type": "array",
            "title": "Risk History"
          },
          "recent_rainfall": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeatherSnapshotSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_sensor_reading": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SensorReadingSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_manual_override": {
            "type": "boolean",
            "title": "Is Manual Override",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "zone",
          "recent_readings",
          "risk_history"
        ],
        "title": "ZoneDetailResponse",
        "description": "Zone detail response schema."
      },
      "ZoneListResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "coordinates": {
            "$ref": "#/components/schemas/CoordinatesSchema"
          },
          "upstream_zone_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Zone Id"
          },
          "latest_assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskAssessmentSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "thresholds": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ZoneThresholdsSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "sensor_health": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensor Health"
          },
          "is_manual_override": {
            "type": "boolean",
            "title": "Is Manual Override",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "coordinates"
        ],
        "title": "ZoneListResponse",
        "description": "Zone list item response schema."
      },
...
      "ZoneThresholdsSchema": {
        "properties": {
          "water_warning_level": {
            "$ref": "#/components/schemas/WaterLevelSchema"
          },
          "water_critical_level": {
            "$ref": "#/components/schemas/WaterLevelSchema"
          },
          "heavy_rain_threshold_mm": {
            "type": "number",
            "title": "Heavy Rain Threshold Mm"
          }
        },
        "type": "object",
        "required": [
          "water_warning_level",
          "water_critical_level",
          "heavy_rain_threshold_mm"
        ],
        "title": "ZoneThresholdsSchema",
        "description": "Zone thresholds schema for HTTP response."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}